storyboard-creation/requirements.txt
openai>=1.40.0
openai-agents
playwright>=1.46.0
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
tenacity>=9.0.0
